Rename input variable classes
